requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.18.0
opencv-python>=4.5.0
matplotlib>=3.5.0
pandas>=1.3.0
tqdm>=4.62.0

# Optional dependencies
napari>=0.4.15  # For interactive visualization
pyyaml>=6.0     # For YAML configuration files
pytest>=7.0.0   # For running tests

# Development dependencies
black>=22.0.0   # Code formatting
flake8>=4.0.0   # Code linting 